requirements.txt
django==1.7.9
Pillow==3.1.1
django-braces==1.8.1
djangorestframework==3.3.3
python-social-auth==0.2.18
facepy==1.0.8
